# ENH: faster region building in spatial.SphericalVoronoi