In the `spatial.SphericalVoronoi` module, regions can be built faster by offloading more of the calculation to numpy.

Benchmarks:
```
============ ============= =============
 num_points      master       this PR
------------ ------------- -------------
     10         546±1μs       494±4μs   
    100         1.69±0ms      1.32±0ms  
    1000      13.6±0.02ms   10.2±0.08ms 
    5000       71.7±0.3ms   53.6±0.07ms 
   10000       148±0.7ms     112±0.6ms  
============ ============= =============
```